+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Major Changes
 
 ### Minor Changes
+- Added the `CoreNamespace.getTokensForOwner()` method to get all the token balances and token metadata for a given address.
 - Added the `wyvern` and `cryptopunks` marketplaces in the `NftSaleMarketplace` enum to the `NftNamespace.getNftSales()` method.
 - Added the `blockHash` field to the `OwnedNftsResponse` returned by `NftNamespace.getNftsForOwner()` to track the block hash that the request was based on.